refactor(stats): move grpc_status_code_to_name to data-pipeline/otlp
