+++ libdd-trace-stats/src/span_concentrator/mod.rs
@@ -11,9 +11,7 @@ use aggregation::StatsBucket;
 
 mod aggregation;
 use aggregation::BorrowedAggregationKey;
-pub use aggregation::{
-    grpc_status_code_to_name, FixedAggregationKey, OtlpExactCell, OtlpExactGroup, OtlpStatsBucket,
-};
+pub use aggregation::{FixedAggregationKey, OtlpExactCell, OtlpExactGroup, OtlpStatsBucket};
 
 pub mod stat_span;
 pub use stat_span::StatSpan;
